fix: movement units (#983)
This release fixes a bug with movement units display.  The default unit is "meters" for traveller but the display showed "ft".  This was due to the list selection being misconfigured.  It is fixed now, but requires a migration to do so. The labeling is also consistent with the 5e system to increase capability with other modules.  Also, two additional movement units have been added, parsecs and grid units (for those who measure distance in squares or hexes).
 ****Important: Please backup your world data before installing this update****
I've tried this on several worlds and it seems fine, but the migration utility is new.