feat: add processor component and allow software on any item (#1301)
This release adds a new attachment type - a "processor".  This attachment adds a computer-like feature to any item accepting attachments. It tracks bandwidth used for any related software that is  attached to the same item. Note that computers items may not use this attachment type. Thanks to @DaishoChikara for the suggestion.
Also included are some style fixes and updated Spanish translations from @ForjaSalvaje.