feat: Add basic weapon range modifier support (#1501)
This release adds basic support for weapons range modifiers (thanks to the suggestion of @zee).  Because this is a complex feature and depends on the rules used, there is a separate wiki page to describe how it works (https://github.com/xdy/twodsix-foundryvtt/wiki/Weapon-Range-Roll-Modifiers).  Please review that page before using this new feature.

Because this is such a big change, there are bound to be some bugs.  So, you might want to try the feature out before running in a live session.  There is also a migration for the "Range Band" field for these using the CE Rules.  You may have to modify that field for existing weapons as the migration can't cover 100% of the cases.  Also, please backup your data before running!

Other fixes: 
* Monks Tile Trigger style fixes
* Add roll description if also AE changes value of characteristic
* Add missing width and height to svg files
* Add species field to animal sheet
* ES localization updates (thanks to @ForjaSalvaje)