fix: Improve scaling changes for Traveller Sheet and add collapsable attachments(#1619)
This update improves the scaling of lists on the Traveller sheet when changing sheet size. Previously, lists were a fixed in width based on the font size. 

Also, attachments and consumables listed with parent item are now collapsable.  Two new settings for the actor sheet set the default expanded state when a sheet is opened.  Thanks to @Henrox for the suggestion.