feat: add basic DM modifier for component damage (#1817)
This release adds a new feature to have an optional DM for ship components when the component has damage.  For each hit, a new setting value under 'Advanced Ruleset Settings -> Ship -> Component Damage DM' is multiplied by the number of hits to add an automatic DM for a ship position action.  For example, if the setting is -1 and the component has two hits, a -2 DM is added to the position roll using that component.  If you don't want to use this automatic DM, leave the setting to its default value of zero.

This release also includes a small change when using Classic Traveller rules.  For characteristic modifiers, these values are displayed as a dash on the sheets (rather than zero) to avoid confusion.

Also, Active Effect tooltips have been refactored to compute them before the sheet loads instead of on-render.